Feat/check user input for emptiness
- Add check user input for emptiness
- Add system prompt init to classifier constructor

+++ internal/app/bot/bot.go
@@ -63,6 +63,10 @@ func (b *Bot) Run() {
 		scanner.Scan()
 		userInput := scanner.Text()
 
+		if userInput == "" {
+			continue
+		}
+
 		if userInput == "exit" {
 			err := syscall.Kill(pid, syscall.SIGINT)
 			if err != nil {
